+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # ChangeLog
 
+--- develop ---
+
+* feature#2: Provide Site and Location filters when viewing UPS's
+
 --- 1.2 ---
 
 * issue: Fix issue with replication full sync errors